chore: fetch 기반 apiFetch 유틸 제거 (#204)
