+++ src/utils/apiError.ts
@@ -1,0 +1,11 @@
+export class ApiError extends Error {
+  status: number
+  detail?: string
+
+  constructor(status: number, message: string, detail?: string) {
+    super(message)
+    this.name = 'ApiError'
+    this.status = status
+    this.detail = detail
+  }
+}
